Lodha Thanisandra Main Road Bengaluru
Lodha Project Off Thanisandra Main Road is a residential project that is designed that come with a premium living lifestyle. Come around the positive vibes where you are offered living spaces that perfectly suit you & your family members.

This top-level project is designed with the principles of Zen where you can attain a life that offers ideal stability. A place that is pampered with luxury and nurtured with nature, coming along with the vibrancy of classic natural life. Justifying the name as it creates a dream garden for your smooth lifestyle.

Lodha Manyata Tech Park Bangalore is a much-awaited project that is close to everything that staples. So, get all set to live a clutter-free and fluid architectural design that is sure to inspire the luxury lifestyle. A home that reflects the mystic haven which is just created for you and your family members. This destination is where life flourishes in serene strides and a sanctuary that does not overwhelm you with excess but does more with easiness.

Lodha Thanisandra Main Road Bengaluru floor plans offer 2 BHK, 3 BHK & 4 BHK residential apartments in the size of various Sq.ft. area. It is a retreat where you revitalize your spirits to soar to the skies and loaded with harmony that enlightens you fully. It is meticulously planned to offer a sense of balance to your lifestyle. From large windows that frame beautiful vistas of sprawling outdoors, to the natural lighting in your living room, you will notice the interpretations of a design attitude.

It welcomes you to an oasis of calm amidst verdant gardens and exquisite lifestyle conveniences where you can attain a fine balance between aesthetics and performance. The well-developed social infrastructures falling at short distances add more worth by creating a charming area. The famous Manyata and L&T Tech parks are just at a distance of 9 km from this apartment. Get the best education facilities for your kids from the reputed schools like Bangalore International and Delhi Public School within the proximity.

For More Details:
Visit Here: https://bit.ly/3xTMAvv